Implement trimming of zero-impact `allow(...)` inner attribute elements
This is, for now, only a subset of the attributes that c2rust might generate.

We don't (yet) bother erasing naming-related ones that are very likely to
have an impact on code translated from C. In particular, we don't try to erase
`non_camel_case_types`, `non_snake_case`, or `non_upper_case_globals`.