Try to be more helpful when upstream `c2rust` fails
Before, we would just print a stack trace, which wasn't of much use.
The stack trace would include the command which failed, but that would reference
freshly-deleted temporary directories, so that was misleading at best.

Now, we print a prominent notice explaining what the failed step was.
We also print the failed command's stdout and stderr, which had previously
been silently swallowed.