Avoid having `chkc` consume standard libraries on Mac
Mac standard libraries use attributes that cause `goblint-cil` to barf.
For now, the quick fix is to use a forward declaration instead of `#include`.

Longer term, we might fork/substitute a different version of CIL, or simply
feed CodeHawk a sanitized copy of the preprocessed code with any
undesirable attributes erased.