Insert `#include "xj_globals.h" before first function that needs it
If we put it at the start of the translation unit, we won't have the necessary
types in scope yet. And if we redefine the types to be in scope at the start,
we'll get type redefinition errors later on.

This placement is not guaranteed to be correct because a C file might put
an #include which provides a type needed by XjGlobals
after a mutable-global-accessing function definition. If that happens,
the only solution AFAICT is to reorder the includes/definitions, which
is itself not trivial to do with high fidelity at the source level.