Identify globals without initializers
I don't see a way to robustly do this via libclang's Python bindings, at least with
the bindings as of LLVM 18. In particular, for a VarDecl of a sized array type,
I don't see any way to distinguish the expr for the size from the initializer.

One could maybe try reflecting on type information but this seems more robust.