Support compile_commands.json with relative paths out of nested directories
This occurs for `wpa_supplicant`, which has a directory structure like this:

```
hostap/                # root directory
   wpa_supplicant/     # target application
   src/                # shared source dir
   ...                 # other targets
```

The `wpa_supplicant/` directory has a Makefile to build the target application,
which involves compiling sources in the adjacent `src/` directory. This results
in compilers being passed relatives paths to files which start with `../src/`.