Reimplement c2rust's cc wrappers
For non-trivial codebases, with multiple libraries and/or executables, we need
more information than just `compile_commands.json`. That file contains all
translation units for all targets, but we need to know which files were used
in which targets -- otherwise, we risk symbol conflicts, etc. The compilation
database also does not list produced shared libraries, or other link-related
entries.

In theory CMake can give us accurate target information via the file query API,
but in practice there are some issues because CMake exposes a dependency-oriented
view of the project. So, for example, targets will list their associated source files, but
a single source file path can be compiled multiple ways with multiple flags, and it's left
to you to reconstruct the right intermediate (`.o`) file. Fragile!

Hopefully this wrapper-based approach is more robust,
at least for CMake via the `CMAKE_C_*_LAUNCHER ` flags.

This is a close re-implementation of `c2rust/scripts/cc-wrappers/`.
The biggest functionality difference is to permit the use of clang
as either compiler-driver or linker-driver.