Change strategy for synthesized compile_commands.json
Instead of having the current codebase as the database's `"directory"`,
this keeps the original build dir and rewrites referenced files.

As a consequence of keeping the original build dir, we change the
build to happen in the results directory, rather than an anonymous
temporary directory, which would get deleted before we're done with it.