Fix filesystem TOCTOU bug in reprovisioning
Reprovisioning of LLVM will delete the xj_prepare_findfnptrdecls directory,
with the intention that builds will regenerate it on demand.
The reprovisioning check only happens on calls to `hermetic.run()` and friends,
which occurs after the check that the directory exists. The result is that the
directory would be deleted in between its check and use.

Doing the Rust build first moves the deletion step before the check.
(If we ever parallelize the build steps, we'll need to reprovision more aggressively.)